requirements.txt
# The web framework for perfectionists with deadlines.
Django==1.10.6

# Image processing library, formerly known as and backwards compatible with PIL. May require extra system dependencies.
Pillow==4.0.0

# http://daringfireball.net/projects/markdown/ - used for parsing markdown in templates.
markdown2==2.3.3

# PostgreSQL database driver for Python. Requires postgresql-server to be installed on the system.
psycopg2==2.7

# Super-fast, widely used memory caching framework. Requires memcached to be installed on the system.
python-memcached==1.58

# Thumbnail library. Requires Pillow and memcached for current implementation.
sorl-thumbnail==12.3

# Framework for building RESTful APIs.
djangorestframework==3.5.4

# Smart filtering/ordering/searching for Django Rest Framework.
django-filter==1.0.1

# Auto-generates API documentation using code introspection.
django-rest-swagger==2.1.1

# Model history, rollback, recover deleted models, etc.
django-reversion==2.0.8

# ORM cache with "automatic granular event-driven invalidation"
django-cacheops==3.1.2

# Django cache backend for Redis
django-redis==4.7.0

